requirements.txt
click>=8.1
rich>=13.0
httpx>=0.27
pydantic>=2.0
thefuzz>=0.22
python-Levenshtein>=0.25
pyyaml>=6.0
streamlit>=1.30
plotly>=5.18
openpyxl>=3.1
pandas>=2.1
pytest>=8.0
